+++ phenodata/dwd/pheno.py
@@ -258,11 +258,6 @@ class DwdPhenoData(object):
         # Load multiple files into single DataFrame
         for path in tqdm(paths, ncols=80):
 
-            # Skip invalid files
-            if 'Kulturpflanze_Ruebe_akt' in path or 'Kulturpflanze_Ruebe_hist' in path:
-                logger.warning('Skipping file "{}" due to invalid header format (all caps)'.format(path))
-                continue
-
             logger.debug('Processing file "{}"'.format(path))
 
             # Acquire DataFrame from CSV data